+++ BusinessObjects/Models/ResourcePool.cs
@@ -50,11 +50,6 @@ namespace forCrowd.WealthEconomy.BusinessObjects
         [DisplayOnEditView(true)]
         public decimal InitialValue { get; set; }
 
-        [Display(Name = "Main Element")]
-        [DisplayOnListView(false)]
-        [DisplayOnEditView(false)]
-        public int? MainElementId { get; set; }
-
         [Display(Name = "Use Fixed Resource Pool Rate")]
         [DisplayOnListView(false)]
         [DisplayOnEditView(true)]
@@ -80,7 +75,6 @@ namespace forCrowd.WealthEconomy.BusinessObjects
         public int? RatingCount { get; private set; }
 
         public virtual User User { get; set; }
-        public virtual Element MainElement { get; set; }
         public virtual ICollection<Element> ElementSet { get; set; }
         public virtual ICollection<UserResourcePool> UserResourcePoolSet { get; set; }
 
@@ -109,27 +103,6 @@ namespace forCrowd.WealthEconomy.BusinessObjects
             user.UserResourcePoolSet.Add(userResourcePool);
             UserResourcePoolSet.Add(userResourcePool);
             return userResourcePool;
-        }
-
-        /// <summary>
-        /// Sets the first element in its list as the main element
-        /// </summary>
-        /// <returns></returns>
-        public ResourcePool SetMainElement()
-        {
-            if (MainElement != null)
-            {
-                // TODO What if there is an existing MainElement?
-            }
-
-            var mainElement = ElementSet.FirstOrDefault();
-            if (mainElement != null)
-            {
-                MainElement = mainElement;
-                mainElement.ResourcePoolMainElementSubSet.Add(this);
-            }
-
-            return this;
         }
 
         #endregion